# chore: faster sentry + use native polyfills (#16025)