--- a/.depcheckrc.yml
+++ b/.depcheckrc.yml
@@ -15,6 +15,8 @@ ignores:
   - 'react-compiler-runtime'
   # This is used on the patch for TokenRatesController of Assets controllers, for we to be able to use the last version of it
   - cockatiel
+  # This is used for the buffer polyfills
+  - '@craftzdog/react-native-buffer'
 
   # Note: Everything below this line should be removed after investigation
   # TODO: Investigate each dependency to see whether it's used