--- a/app/util/test/testSetup.js
+++ b/app/util/test/testSetup.js
@@ -6,8 +6,13 @@ import { mockTheme } from '../theme';
 import Adapter from 'enzyme-adapter-react-16';
 import Enzyme from 'enzyme';
 import '@shopify/flash-list/jestSetup';
+import base64js from 'base64-js';
 
 Enzyme.configure({ adapter: new Adapter() });
+
+// Set up global polyfills for base64 functions
+global.base64FromArrayBuffer = base64js.fromByteArray;
+global.base64ToArrayBuffer = base64js.toByteArray;
 
 // Mock the redux-devtools-expo-dev-plugin module
 jest.mock('redux-devtools-expo-dev-plugin', () => {});