… ios/Podfile.lock
… package.json
… yarn.lock