<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been completely filled out, and PR status checks have passed at least once. -->

## **Description**

This is a PR from @Nodonisko at Margelo to make Sentry faster. Original PR - https://github.com/MetaMask/metamask-mobile/pull/16025

This PR could speed up Sentry events processing around up to ~15x. I am not able to consistently reproduce slow Sentry (it depends on how much breadcrumbs and extra data are collected), but performance gain helps even when sending relatively small amounts of extra data. For small event time needed dropped from 100ms to under <10ms. For big events I randomly encountered it dropped from 1s to <100ms.

In future this patch for Sentry might not be necessary, because Sentry team is working on fix in library itself, but it might take some time.

There is also second part of PR modifying `metro.config.js` to use faster native polyfills, this could speed up app in some other places.

## **Related issues**

Fixes:
[#15560](https://github.com/MetaMask/metamask-mobile/issues/15560)

## **Manual testing steps**

I would recommend to run smoke test across whole app because this polyfilled libraries could be used in many places eveywhere in the app.

I would also recommend to verify that events from Sentry are received in Sentry dashboard.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the before and after of your change. -->

### **Before**

Sending big Sentry event before changes on emulator takes around 380ms. For real slow device it's around 800ms.

<img width="723" alt="Screenshot 2025-06-03 at 16 02 24" src="https://github.com/user-attachments/assets/c253843e-6c98-447f-89ca-fe82b92af817" />

For small events it's around 100ms on real device.

### **After**

After changes it takes around 40ms to send big event and around 80ms on real device.

<img width="618" alt="Screenshot 2025-06-03 at 16 00 27" src="https://github.com/user-attachments/assets/53f955d5-fd67-40f2-93ff-ded97dad2f4e" />


For small events on real device it's around 10ms.

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding
Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [x] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.

<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been completely filled out, and PR status checks have passed at least once.
-->

## **Description**

<!--
Write a short description of the changes included in this pull request, also include relevant motivation and context. Have in mind the following questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [ ] I’ve followed [MetaMask Contributor Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask Mobile Coding Standards](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling guidelines](https://github.com/MetaMask/metamask-mobile/blob/main/.github/guidelines/LABELING_GUIDELINES.md)). Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described in the ticket it closes and includes the necessary testing evidence such as recordings and or screenshots.